fix: move HandlesConditionsHarness to PSR-4 file for PHPCS
PHPCS required filename to match class name; extract harness into
Tests\Unit\Traits\HandlesConditionsHarness so duster quality checks pass.
